backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15.0,<4.0
django-cors-headers>=4.3.0,<5.0
psycopg2-binary>=2.9.9,<3.0
boto3>=1.34.0,<2.0
pillow